# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentation for the gdsi API",
        "title": "Gdsi API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/v1/auth/signup": {
            "post": {
                "description": "Signup a new account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Signup",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.SignupRequestModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.TokensResponseModel"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/failure.ValidationFailure"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/failure.Failure"
                        }
                    }
                }
            }
        },
        "/v1/auth/tokens/access": {
            "post": {
                "description": "Login and get a new access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginRequestModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.TokensResponseModel"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/failure.ValidationFailure"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/failure.Failure"
                        }
                    }
                }
            }
        },
        "/v1/auth/tokens/refresh": {
            "post": {
                "description": "Get a refreshed access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh token",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshTokenRequestModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.TokensResponseModel"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/failure.Failure"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.LoginRequestModel": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "auth.RefreshTokenRequestModel": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "auth.SignupRequestModel": {
            "type": "object",
            "properties": {
                "dob": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "auth.TokensResponseModel": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "failure.Failure": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "failure.InvalidField": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "location": {
                    "description": "location specifies where the field is conmming from (path, query, body)",
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "failure.ValidationFailure": {
            "type": "object",
            "properties": {
                "invalid_fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/failure.InvalidField"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter the Bearer token in the format: Bearer token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}